# Generalise monads used by packet streams