Network.Pcap.Streaming.offline: Generalise to MonadResource
Network.Pcap.Streaming.online: Generalise to MonadIO