@fosskers 2018-03-12
Generalizing in this way unfortunately introduces a performance degredation:

#### Before `MonadFoo` constraints

```
benchmarking Benchmarks/Count all entries in stream
time                 4.189 ms   (4.145 ms .. 4.220 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 4.230 ms   (4.191 ms .. 4.305 ms)
std dev              167.1 μs   (93.54 μs .. 298.2 μs)
variance introduced by outliers: 22% (moderately inflated)

benchmarking Benchmarks/Sum all timestamps
time                 4.390 ms   (4.335 ms .. 4.433 ms)
                     0.998 R²   (0.996 R² .. 1.000 R²)
mean                 4.579 ms   (4.505 ms .. 4.700 ms)
std dev              297.8 μs   (182.1 μs .. 453.8 μs)
variance introduced by outliers: 40% (moderately inflated)
```

#### After `MonadFoo` contraints

```
benchmarking Benchmarks/Count all entries in stream
time                 4.596 ms   (4.572 ms .. 4.626 ms)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 4.602 ms   (4.573 ms .. 4.652 ms)
std dev              107.3 μs   (64.28 μs .. 181.1 μs)

benchmarking Benchmarks/Sum all timestamps
time                 4.807 ms   (4.766 ms .. 4.844 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 4.830 ms   (4.811 ms .. 4.857 ms)
std dev              72.12 μs   (60.82 μs .. 86.14 μs)
```

which is something I've experienced before in code that uses `streaming`. Is it possible to `lift` manually in client code that's using something heavier than raw `ResourceT IO`?