@endgame 2018-03-12
Answering the question as asked: `Stream` has instances of the `mmorph` type classes, so it's possible to use `hoist` and friends, so I think the fixed base monad is not a deal-breaker.

That said, what if you added `SPECIALIZE` pragmas for the simplest concrete instances?

```haskell
{-# SPECIALIZE offline :: FilePath -> Stream (Of Packet) (ResourceT IO) () #-}
{-# SPECIALIZE online :: String -> Int -> Bool -> Int64 -> Stream (Of Packet) IO () #-}
-- And maybe
{-# SPECIALIZE packets :: PcapHandle -> Stream (Of Packet) IO () #-}
```