@fosskers 2018-03-12
Yup, that brings the numbers back down. Thanks for the suggestion, I can't believe I forgot about `SPECIALIZE`.